added -Wall and -Wextra and fixed the warnings produced

+++ irc.cpp
@@ -37,7 +37,7 @@ int irc::connect(string hostname, string port)
 	}
 	cout << "connected to " << hostname << ":" << port << endl;
 	connected = true;
-	for(int i=0; i<channels.size(); ++i)
+	for(size_t i=0; i<channels.size(); ++i)
 	{
 		server->sendMsg("JOIN " + channels.at(i));
 		string message;
@@ -67,7 +67,7 @@ int irc::joinChannel(string channel)
 
 int irc::partChannel(string channel)
 {
-	for(int i=0; i<channels.size(); ++i)
+	for(size_t i=0; i<channels.size(); ++i)
 	{
 		if(channels.at(i) == channel)
 		channels.push_back(channel);
